+++ src/toolkit/domain_services/cloudflare.rs
@@ -4,23 +4,14 @@ use serde::{Deserialize, Serialize};
 // https://developers.cloudflare.com/api/operations/dns-records-for-a-zone-update-dns-record
 #[derive(Clone, Serialize, Deserialize, Debug)]
 pub struct Cloudflare {
-    pub email: String,
-    pub zone_id: String,
-    pub dns_record_id: String,
     pub api_token: String,
-    pub name: String,
-    pub r#type: String,
-    pub proxied: Option<bool>,
     pub comment: Option<String>,
+    pub dns_record_id: String,
+    pub email: String,
+    pub name: String,
+    pub proxied: Option<bool>,
+    pub r#type: String,
     pub tags: Option<Vec<String>>,
     pub ttl: Option<usize>,
-}
-
-#[cfg(feature = "dyndns2")]
-#[derive(Clone, Serialize, Deserialize, Debug)]
-pub struct Dyndns2 {
-    pub service_uri: String,
-    pub hostname: String,
-    pub username: String,
-    pub password: String,
+    pub zone_id: String,
 }
